requirements.txt
sentence-transformers
faiss-cpu
openai
numpy
